Add support for custom yosys templates
Edalize creates a template for launching synthesis with yosys, but
there are cases when specific commands need to be injected that is
not possible with the current solution. This patch allows overriding
the default template file with a user-specified alternative byt setting
the 'yosys_template' tool option in the backend. The user template can
still source edalize_yosys_procs.tcl and get access to file lists,
parameters and such

FuseSoC users are recommended to specify the template file with the 'user'
file_type and use 'copyto' to make the file available in a known location
for yosys.